Updated dependencies. Added google.guava dependency in order to safely find Controllers.

+++ studio-api/src/main/java/cdy/studioapi/config/AuthorityPersister.java
@@ -1,8 +1,8 @@
 package cdy.studioapi.config;
 
-import cdy.studioapi.StudioApi;
 import cdy.studioapi.infrastructure.PermissionRepository;
 import cdy.studioapi.models.Permission;
+import com.google.common.reflect.ClassPath;
 import lombok.AllArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
@@ -11,8 +11,9 @@ import org.springframework.lang.NonNull;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Component;
 
-import java.io.File;
+import java.io.IOException;
 import java.util.*;
+import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
@@ -26,49 +27,52 @@ public class AuthorityPersister implements ApplicationListener<ApplicationReadyE
     @Override
     public void onApplicationEvent(@NonNull ApplicationReadyEvent event) {
         var persistedPermissions = permissionRepository.findAll();
-        var controllerMethodPermissions = getPermissions(getControllers());
+        var controllers = getControllers();
+        var controllerMethodPermissions = getPermissions(controllers);
 
         // Find the difference between the permissions in the database and the permissions in the controllers
         var newPermissions = controllerMethodPermissions.stream()
                 .filter(p -> persistedPermissions.stream().noneMatch(pp -> p.getName().equals(pp.getName())))
                 .toList();
 
+        if (!newPermissions.isEmpty())
+            logger.log(Level.INFO, "New permissions found: {0}", newPermissions.stream().map(Permission::getName).collect(Collectors.joining(", ")));
+
         // Find the permissions that are not used in the controllers
         var unusedPermissions = persistedPermissions.stream()
                 .filter(p -> controllerMethodPermissions.stream().noneMatch(pp -> p.getName().equals(pp.getName())))
                 .toList();
+        if (!unusedPermissions.isEmpty())
+            logger.log(Level.INFO, "Unused permissions found: {0}", unusedPermissions.stream().map(Permission::getName).collect(Collectors.joining(", ")));
 
-        // Persist the new permissions
-        permissionRepository.saveAllAndFlush(newPermissions);
 
-        // Delete the unused permissions
-        permissionRepository.deleteAll(unusedPermissions);
+        if (!newPermissions.isEmpty()) {
+            // Persist the new permissions
+            permissionRepository.saveAllAndFlush(newPermissions);
+            logger.log(Level.INFO, "{0} new permissions persisted.", newPermissions.size());
+        }
+
+        if (!unusedPermissions.isEmpty()) {
+            // Delete the unused permissions
+            permissionRepository.deleteAll(unusedPermissions);
+            logger.log(Level.INFO, "{0} unused permissions deleted.", unusedPermissions.size());
+        }
     }
 
     // TODO: Dosya yolunda eğer boşluk varsa hata fırlatıyor başka karakterlerde de aynı sorun çıkabilir
     // TODO: Daha iyi bir yöntem bul!
-    private static List<Class<?>> getControllers() {
-        ClassLoader classLoader = StudioApi.class.getClassLoader();
-        String packageName = "cdy.studioapi.controllers";
-        String packagePath = packageName.replace('.', '/');
-
-        List<Class<?>> classes = new ArrayList<>();
-        File packageDir = new File(Objects.requireNonNull(classLoader.getResource(packagePath)).getFile());
-
-        for (File file : Objects.requireNonNull(packageDir.listFiles())) {
-            if (file.isFile() && file.getName().endsWith(".class")) {
-                String className = packageName + '.' + file.getName().substring(0, file.getName().length() - 6);
-                Class<?> clazz = null;
-                try {
-                    clazz = classLoader.loadClass(className);
-                } catch (ClassNotFoundException e) {
-                    logger.log(java.util.logging.Level.SEVERE, e.getMessage(), e);
-                }
-                classes.add(clazz);
-            }
+    private static Set<Class<?>> getControllers() {
+        try {
+            return ClassPath.from(ClassLoader.getSystemClassLoader())
+                    .getAllClasses()
+                    .stream()
+                    .filter(clazz -> clazz.getPackageName()
+                            .equalsIgnoreCase("cdy.studioapi.controllers"))
+                    .map(ClassPath.ClassInfo::load)
+                    .collect(Collectors.toSet());
+        } catch (IOException e) {
+            throw new IllegalArgumentException("Error while loading Controllers", e);
         }
-
-        return classes;
     }
 
     private static List<String> getMethodAuthorities(String query) {
@@ -83,7 +87,7 @@ public class AuthorityPersister implements ApplicationListener<ApplicationReadyE
         return list;
     }
 
-    private static List<Permission> getPermissions(List<Class<?>> classes) {
+    private static List<Permission> getPermissions(Set<Class<?>> classes) {
         var permissions = classes.stream()
                 .flatMap(c -> Arrays.stream(c.getDeclaredMethods()))
                 .filter(m -> m.isAnnotationPresent(PreAuthorize.class))
